Update patch set 2
Patch Set 2:

When I tried to update the bootloader, I had to download the stock update and extract the file from the archive. If we make the new bootloader mandatory, then there needs to be a simple download either on XDA (Official CM-12.1 thread) or in the CyanogenMod wiki. Otherwise I assume it's too difficult for many people to find that bootloader. (Even better would be a zip-file that can be flashed to update the bootloader.)

Patch-set: 2