Update patch set 1
Patch Set 1:

The log is needed in case there are buttons that are not detected. All the other messages are disabled and must be enabled at compile time, but I'm mostly interested in buttons that are not recognized.

Anyway, Google says that the 110-180 ohm should be used for voice assist and the 360-680 ohm range should be used for volume down. The two ranges are pretty far from each other, so I can say that his headset doesn't follow the specification (unless I made some mistake in the code) and I can't fix that.

However, he can easily get around the problem, he just needs to remap the button. All he has to do is to edit /system/usr/keylayout/msm8226-tapan9302-snd-card_Button_Jack.kl and replace

  key 217   VOICE_ASSIST

with

  key 217   VOLUME_DOWN

Patch-set: 1