Update patch set 1
Patch Set 1:

The fix was specific to titan and thea, the problem was their kernel driver/proximity sensor. Ours is working fine.

I've recently noticed a different problem on falcon/peregrine and currently testing a patch: http://review.cyanogenmod.org/#/c/137447/. Probably it's the same problem you are referring to.

The main problem is that I don't know how to reproduce it and it happened only twice, so I couldn't really debug it. All I know is that it's related to the significant motion sensor/accelerometer, which prevents the device from entering low power states. Since nothing changed in a really long time kernel side, I'd say the bug had always been there, but it's happening only now because of some other change.

Patch-set: 1