Update patch set 1
Patch Set 1:

I doubt the proximity check on wake is the cause of the issue. It may contribute, but I think the real problem is somewhere else.

The last time the bug happened, I looked at the dmesg and I found http://pastebin.com/raw/nV8qyd4W repeated over and over. That basically says that there's a wakeup IRQ (lis3dh_irq1, i.e. sig motion sensor) that is waiting to be processed. Since this is repeated over and over, I suspect that what is happening is exactly what's described in the commit message of http://review.cyanogenmod.org/#/c/137746/.

This should happen only after the device entered Doze mode. I'm not saying that it happens everytime the device enters Doze mode, it's just a requirement.

Patch-set: 1