Update patch set 1
Patch Set 1:

Just to inform that even with these patches, battery life is pretty bad on peregrine. Reverting to stock sensors hal solves it. 

"lis3dh_irq1 pending, suspend aborted" appears to be fixed, but the device is still not going to deep sleep most of the times. If you need help tracking down the cause and testing stuff, just drop me a message. Keep up the good work

Patch-set: 1