Update patch set 1
Patch Set 1:

Everything looks fine. The device enters deep sleep and whenever it wakes up, the reason is not related to the sensors, except for one wakeup, caused by the significant motion sensor. This happend 6 seconds before you turned the screen on, so I assume there actually was a significant motion.

Extracting the timestamps of the various "PM: suspend {entry,exit}" lines, I could see that yes, the device wakes up every now and then, but when it does, it stays awake for hundreds of milliseconds and then goes back to sleep.

Let's see how it goes without the significant motion sensor. The device will never enter Doze, but the accelerometer will never be enabled while the screen is off. The power consumption of the accelerometer configured for motion detection should really be negligible, but maybe I'm wrong.

Patch-set: 1